[R5] Move each entity dto for their entity package
