+++ src/main/java/com/mirna/hospitalmanagementapi/domain/dtos/doctor/DoctorDTO.java
@@ -1,5 +1,6 @@
-package com.mirna.hospitalmanagementapi.domain.dtos;
+package com.mirna.hospitalmanagementapi.domain.dtos.doctor;
 
+import com.mirna.hospitalmanagementapi.domain.dtos.AddressDTO;
 import com.mirna.hospitalmanagementapi.domain.enums.Specialty;
 
 import jakarta.validation.Valid;
