+++ src/main/java/com/mirna/hospitalmanagementapi/domain/dtos/patient/PatientDTO.java
@@ -1,4 +1,6 @@
-package com.mirna.hospitalmanagementapi.domain.dtos;
+package com.mirna.hospitalmanagementapi.domain.dtos.patient;
+
+import com.mirna.hospitalmanagementapi.domain.dtos.AddressDTO;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
