+++ src/main/java/com/mirna/hospitalmanagementapi/domain/entities/Address.java
@@ -1,7 +1,7 @@
 package com.mirna.hospitalmanagementapi.domain.entities;
 
 import com.mirna.hospitalmanagementapi.domain.dtos.AddressDTO;
-import com.mirna.hospitalmanagementapi.domain.dtos.DoctorDTO;
+import com.mirna.hospitalmanagementapi.domain.dtos.doctor.DoctorDTO;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
