+++ src/main/java/com/mirna/hospitalmanagementapi/domain/entities/Doctor.java
@@ -1,6 +1,6 @@
 package com.mirna.hospitalmanagementapi.domain.entities;
 
-import com.mirna.hospitalmanagementapi.domain.dtos.DoctorDTO;
+import com.mirna.hospitalmanagementapi.domain.dtos.doctor.DoctorDTO;
 import com.mirna.hospitalmanagementapi.domain.enums.Specialty;
 
 import jakarta.persistence.Column;
