+++ src/test/java/com/mirna/hospitalmanagementapi/unit/application/controllers/DoctorControllerTest.java
@@ -17,8 +17,8 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mirna.hospitalmanagementapi.domain.dtos.AddressDTO;
-import com.mirna.hospitalmanagementapi.domain.dtos.DoctorDTO;
-import com.mirna.hospitalmanagementapi.domain.dtos.DoctorUpdatedDataDTO;
+import com.mirna.hospitalmanagementapi.domain.dtos.doctor.DoctorDTO;
+import com.mirna.hospitalmanagementapi.domain.dtos.doctor.DoctorUpdatedDataDTO;
 import com.mirna.hospitalmanagementapi.domain.entities.Doctor;
 import com.mirna.hospitalmanagementapi.domain.enums.Specialty;
 import com.mirna.hospitalmanagementapi.domain.repositories.DoctorRepository;
